fix audio channel leaving error

+++ src/main/java/core/InternalEventListener.java
@@ -9,7 +9,9 @@ import net.dv8tion.jda.api.OnlineStatus;
 import net.dv8tion.jda.api.entities.Activity;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
+import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
 import net.dv8tion.jda.api.entities.channel.middleman.AudioChannel;
+import net.dv8tion.jda.api.entities.channel.unions.AudioChannelUnion;
 import net.dv8tion.jda.api.events.guild.voice.GuildVoiceUpdateEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
@@ -19,6 +21,7 @@ import net.dv8tion.jda.api.events.session.ReadyEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
+import net.dv8tion.jda.api.managers.AudioManager;
 import net.dv8tion.jda.api.managers.Presence;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
@@ -116,12 +119,20 @@ public class InternalEventListener extends ListenerAdapter {
                 this.logger.debug("All participants left the voice channel. Clearing the queue.");
                 MusicBox musicBox = Service.GetMusicBoxByGuildId(guild.getId());
                 MusicStreamer streamer = musicBox.getStreamer();
+                AudioManager audioManager = guild.getAudioManager();
+                AudioChannelUnion connectedChannel = audioManager.getConnectedChannel();
+                if (connectedChannel == null) return;
+                VoiceChannel voiceChannel = connectedChannel.asVoiceChannel();
+
                 streamer.clearTracksOfQueue();
                 musicBox.updateEmbed();
-                musicBox.getAudioManager().closeAudioConnection();
-                musicBox.getMusicChannel().sendMessage("모든 참가자가 음성 채널을 나갔습니다. 음악 재생을 종료합니다.").queue(sentMessage -> {
-                    sentMessage.delete().queueAfter(5, java.util.concurrent.TimeUnit.SECONDS);
-                });
+                if (voiceChannel.getId().equals(audioChannel.getId())) {
+                    voiceChannel.getGuild().getAudioManager().closeAudioConnection();
+                    audioManager.closeAudioConnection();
+                    musicBox.getMusicChannel().sendMessage("모든 참가자가 음성 채널을 나갔습니다. 음악 재생을 종료합니다.").queue(sentMessage -> {
+                        sentMessage.delete().queueAfter(5, java.util.concurrent.TimeUnit.SECONDS);
+                    });
+                }
             } catch (GuildManagerNotFoundException ex) {
                 logger.error("GuildManagerNotFoundException occurred while trying to get MusicBox.");
                 guild.getAudioManager().closeAudioConnection();
